produto_service/requirements.txt
apispec==6.8.2
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
flask-smorest==0.46.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
packaging==25.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
typing_extensions==4.14.1
webargs==8.7.0
Werkzeug==3.1.3
